feat(landing): show landing page on connection
Some servers like people to select their character on connection, because of this a landing page can be enabled. This will set the players routing bucket to be different from the base bucket so they do not see oddities from general game play. Once they have selected a character they will be set to the default routing bucket as specified in the fxmanifest meta data.